Initial Commit
This is the first commit on my http server project. My aim is to build a small api without any external dependencies / modules in the hope that I will understand what is happening at a deeper level. I'm also planning on using nvim/tmux to write this whole thing. Currently running app.py will cause the program to bind to localhost:8080 as all good things do. The program returns a simple html response and it is intelligent enough to discern the type of request that it received